fix clippy

+++ crates/primitives/src/transaction/alt_fee.rs
@@ -308,7 +308,7 @@ impl RlpEcdsaEncodableTx for TxAltFee {
 }
 
 impl RlpEcdsaDecodableTx for TxAltFee {
-    const DEFAULT_TX_TYPE: u8 = { Self::tx_type() as u8 };
+    const DEFAULT_TX_TYPE: u8 = { Self::tx_type() };
 
     /// Decodes the inner [TxEip1559] fields from RLP bytes.
     fn rlp_decode_fields(buf: &mut &[u8]) -> alloy_rlp::Result<Self> {
@@ -322,7 +322,7 @@ impl SignableTransaction<Signature> for TxAltFee {
     }
 
     fn encode_for_signing(&self, out: &mut dyn alloy_rlp::BufMut) {
-        out.put_u8(Self::tx_type() as u8);
+        out.put_u8(Self::tx_type());
         self.encode(out)
     }
