+++ crates/primitives/src/transaction/l1_transaction.rs
@@ -241,7 +241,7 @@ impl RlpEcdsaEncodableTx for TxL1Msg {
 }
 
 impl RlpEcdsaDecodableTx for TxL1Msg {
-    const DEFAULT_TX_TYPE: u8 = { Self::tx_type() as u8 };
+    const DEFAULT_TX_TYPE: u8 = { Self::tx_type() };
 
     /// Decodes the inner [TxEip1559] fields from RLP bytes.
     fn rlp_decode_fields(buf: &mut &[u8]) -> alloy_rlp::Result<Self> {
@@ -253,7 +253,7 @@ impl SignableTransaction<Signature> for TxL1Msg {
     fn set_chain_id(&mut self, _chain_id: ChainId) {}
 
     fn encode_for_signing(&self, out: &mut dyn alloy_rlp::BufMut) {
-        out.put_u8(Self::tx_type() as u8);
+        out.put_u8(Self::tx_type());
         self.encode(out)
     }
